Started off the Project
I am pushing all of the code in case of a fault on my local end and for an explanation of what the project is about.

I have copied my getBattery code from the other project in understanding how to use the Tello drone. The main.py for now does a simple route to one area. More will come soon.